+++ VERSION_HISTORY.markdown
@@ -2,6 +2,7 @@
 
 ## 0.2.4
 
+* Fixes #19 - `punic checkout` now "prepares" the project checkout in the same way `punic build` will.
 * Fixes #25 - punic would be tricked into trying to build static frameworks (e.g. Facebook's Bolts framework). This would break the dSYM generation code.
 
 ## 0.2.3